move GetFromApi metods to Services
