+++ src/services/SearchByKeyWord.jsx
@@ -1,4 +1,4 @@
-import axios from '../../services/AxiosConfig'
+import axios from './AxiosConfig'
 
 ////query - find by city/region/counry 
 const SearchByKeyWord = ({query = null, loc = null}) =>{
